Updated openstack/openstack Project: openstack/neutron-lbaas  d6a9396df41b82c66d6f5abba7473f3c9d81bf41
Fix order of arguments in assertEqual

assertEqual expects that the arguments provided to it should be
(expected, observed).If a particluar order is kept as a convention,
then it helps to provide a cleaner message to the developer if Unit Tests
fail. There are several Unit Test files where the arguments for assertEqual
have been swapped. This commit fixes the issue in lbaas.

Change-Id: I890deda05a26c9f84d7e2884970d051faa30830c
Partial-Bug: #1259292
Related-Bug: #1259292