Updated openstack/openstack Project: openstack-infra/lodgeit  f79c65815a3eeb5130040051fdb38a3428a82e51
Fix up the tests and add tox configuration to run them

The alfajor HTTP test wrapper is not working properly.  All the tests
are failing with:

  AttributeError: '_APIClientResponse' object has no attribute '_status_code'

It doesn't look like alfajor is being maintained any more [1, 2], so
remove the dependency on it and run the tests directly through the
werkzeug client instead.

The test_json_get_styles test still failed because the content and
ordering of the styles returned from the API is dependent on the
platform on which the tests run.  Instead of testing against a fixed
expected result, which could be different on another platform,
generate the expected results dynamically.

Update the tox configuration and test-requirements.txt to be able to
run the tests via tox.

[1] https://github.com/idealist/Alfajor
[2] https://github.com/idealistdev/alfajor

Change-Id: I4db6783b9d725ce096575035ea27d28be8e6fed5