Updated openstack/openstack Project: openstack/trove-integration  ff2fafc0767c07b280b637bc609e6d8487f45d72
Use current branch for devstack/projects

This changes defaults the behaviour to use the current branch of
devstack and each project, unless the corresponding environment
variable is set.

When support for variables such as TROVE_BRANCH was added, an
unintended side-effect was introduced.  In the use case of
testing a changeset by creating a new branch for the checkout,
redstack would switch the repo back to master, if the
TROVE_BRANCH variable was not explicitly set.  This forces the
developer/tester to set the variable - if they forget they will
end up inadvertently testing the master branch.

For example, doing the following would not give the
expected result:

cd /opt/stack/trove
git fetch ssh://<fetch_cmd> && git checkout FETCH_HEAD
git checkout -b new_branch
cd ~/trove-integration/scripts
./redstack install

Redstack in this case would switch the repo back to master,
so you would need to remember to add 'TROVE_BRANCH=newbranch'
before the install command.

This fix makes sure that the repo is only switched if one of the
variables is actually set.

Change-Id: I4e74f7c7222548e4e2ec8ce02d29a610d2f2e1da