Updated openstack/openstack Project: openstack/python-glanceclient  b88b510d3b6b497ba315f2e1c1d31c4ee9662912
print usage when no argument is specified for python3

When running just 'glance' under python3, we will get the error:
ERROR: 'Namespace' object has no attribute 'func'

This is because map() is used to decode sys.argv, but under Python3
it returns a map object which is an iterable. Some code later tries
to use this in a boolean context and it will always return True,
even if it's empty.

Change-Id: I2f03e462cb813833b75b9f2de7badd10b10cddff
Closes-Bug: #1295356