Updated openstack/openstack Project: openstack/neutron  91c56a1598cec70f9a612e31b79d79670e8652a2
Fix link in devref guide

warnerrors = True no longer works with Sphinx < 1.3, and this
makes warnings fail silently [2].

A patch [1] is out to fix that, and until pbr releases we will
have to be careful not to bitrot.

This patch fixes the outstanding link error and tweaks the wording
of the affected pill of wisdom.

[1] I680b448471e687919d202e8f2abe57f8ba3b22ee
[2] bug #1496882

Change-Id: Id9cce7ae78c51dceeb56c6b5b8b04340c5378748