Updated openstack/openstack Project: openstack/neutron  1cc051fffe43774ac408c60772b77158544e2425
Tweak RFE guidelines

Liberty is behind us. The RFE process has been in place for a cycle
and some refinements have been necessary to make it more streamlined.

The cut-over section has outlived its purpose so it can be removed.
In its place, this patch clarifies how we are going to track RFE in
Launchpad, once the RFE has been 'approved'.

The tl;dr version of this patch is: from Mitaka onwards, feature
submission must happen via RFE bug report submission. Launchpad
blueprints are going to become a tool to be used at discretion of
the project release manager for milestone tracking purposes.

Change-Id: I370e3397682d9f4ebca215b6711f23a2a4fc4233