Updated openstack/openstack Project: openstack/api-site  84e86566e2b15d0b42d49d73bcea8dd0387fd890
Stop deleting draft firstapp docs

We have a number of firstapp documents in a range of SDKs
that are in draft status.

Particularly, for SDKs like shade which have all four of the
conditional sections complete, we would like to issue a call
for testing, leading to their eventual publication.

Having an easily accessible link to give testers is a crucial
part of that process. This patch removes the lines that delete
the draft firstapp docs so they can be found on the draft site.

Change-Id: I0457babc095ab152512751b7cb453dc5c1957282