Updated openstack/openstack Project: openstack/glance  a170accaa931f44906140241f04376a2157292c1
Decrease test failure if second changes during run

test_create_with_live_time fails gating constantly. This is due to
the logic in the test which breaks if the second changes between the
two timestamps expires_at and updated_at are created. This change 
drops the failure rate to those occasions where we hit that second
change also to minute change.

Closes-Bug: #1505675

Change-Id: I234d80388edd2a659a2374a29179ce23b23bb99a