Updated openstack/openstack Project: openstack/neutron  201ae16eec2964d56cda4cc6ab059ebbd1a422d0
Fix inconsistency in DHCPv6 hosts and options generation

The DHCP agent is inconsistent in how it handles subnets whose
ipv6_address_mode is not slaac.  While the DHCP agent writes out both
DHCPv6 host entries and DHCPv6 options for ports scoped by the subnet
for dnsmasq to use, subnet specific options are not written.

This patch addresses this inconsistency by generating subnet specific
options when the subnet's ipv6_address_mode is not slaac.

Closes-Bug: #1501451

Change-Id: I372c93fa5667e955c7b81181daa2e66f47d85fdf