Updated openstack/openstack Project: openstack/trove  03674b10d96a0ffef080415bc72569ef8429149e
oslo.utils 2.6.0 causing tox test to hang

Forcing the change in requirements.txt is not feasible as all
entries there must match the global-requirements.txt file.
This forces a specific version of oslo.utils *before* the
requirements are loaded, thus ensuring that 2.6.0 is *not* used.

This should be considered a temporary fix only.

Change-Id: I4363c4b8b86efb15fd95d07ce6fff374024b3616
Partial-Bug: #1506402
Co-Authored-By: Amrith Kumar <amrith@tesora.com>