Updated openstack/openstack Project: openstack/cinder  59c5eab6cfcbb1969d3ae660dc488fd0be40e5c3
Use lvm_conf_file directory for LVM_SYSTEM_DIR value.

Currently when LVM_SYSTEM_DIR is added to LVM.LVM_CMD_PREFIX it is
hardcoded as /etc/cinder.

However the configuration option lvm_conf_file could specify a file
located in an arbitrary directory, not just /etc/cinder, so we should
be using the path to the directory containing the specified lvm.conf
file as the value for LVM_SYSTEM_DIR.

Change-Id: I702fdfb3640ef99255fda0fda4d39d0a80ce9dcc
Closes-Bug: #1498651