Updated openstack/openstack Project: openstack/ironic  95c21de106b2b7479dc5163d46cdfdcfcbb73328
Remove functions: _cleanse_dict & format_message

The functions _cleanse_dict & format_message are not used.

_cleanse_dict was used previously by another function that was removed
because it was unused.

format_message usage has been removed as the correct way to get a
message from an exception is six.text_type(exception). Remove it to
prevent people from trying to use it.

Closes-Bug: #1505080
Change-Id: I1a1284b0526787a62b6765bf2e3547dbba822352
Depends-On: I75757b5bd7d9e75ca37d1b9a1efff810ba84b5df