Updated openstack/openstack Project: openstack/api-site  cd14d8cc0ee2b1c67f06a42c7e5472216859bd6c
The servers remote console section is missing RDP and Serial console

There are four consoles in the Compute API, but the doc is missing RDP
and Serial console. This patch adds them.

(df) Updated the descriptions and also added in the missing
"get console log output" call.
(df) Consolidated all server actions and ensured that the
actions match what's in the generated code samples.

Closes-Bug: #1446660

Change-Id: Ibfe994a3be38da2cb53e3edb0df41511a29368cc