Updated openstack/openstack Project: openstack-infra/yaml2ical  f02d892ea6327640ef5079f1af145bf8f8154d6d
Add one off events for skipped meetings

Rather than silently removing the skipped meeting from the calendar as
is the default behavious of exdate add a one off calendar entry for
each skipped meeting.

The aim is to show that the slot is generally consumed just that the
meeting has been explictly cancelled/skipped.

Change-Id: Idd9eb1a04fe9c9661e87978c0128a292bb8e89a4