Updated openstack/openstack Project: openstack/manila  f1ada3fe97942931c181ffe95df6233e3f3c9277
Fix share type model scalability for get request

"share type list" operation greatly depends on amount of records in
"shares" table. It leads to big slowness of just requesting list
of several records of share types. That can take dozens of seconds.

Main reason for it lays in usage of redundant "relationship" of "share
types" table to "shares" table. It is not used in Manila, so remove it.

Change-Id: I463b12e2eb7249b42835be063cd953b56ee7376d
Closes-Bug: #1506240