Updated openstack/openstack Project: openstack/neutron  8a0cace74a87f55c475e76e8b6102374d91fa7d8
Move retries out of ML2 plugin

There are several DB retry decorators in the ML2 plugin.
We want to keep them in the API layer as much possible so
this patch removes all but 1 and adds the retryrequest catch
to the API-level decorator.

The remaining one is the update_port_status which is an internal
RPC called method so it's not clear if there is a benefit to
moving it to the ml2 rpc file.

Change-Id: Ifffd424b693f57b5dd3de4dc625b50acc23abe0f
Closes-Bug: #1505908