Updated openstack/openstack Project: openstack-infra/project-config  10119113831bf0b6a552ef977e3b790d1eb51a48
Avoid name collisions with tripleo-ci jobs

The tripleo-ci jobs were using a regex that just looked for
*-tripleo* in the job name, which also matches tripleo-common and
python-tripleoclient.  This causes the py34 jobs for those projects
to sometimes get scheduled to precise nodes that don't have a py34
interpreter installed.

This changes the tripleo-ci jobs to be named tripleo-ci-* and
updates the regex appropriately so it will no longer match non-toci
jobs.

Change-Id: Ie4d803574221b936b249673436db36fd75940587