Updated openstack/openstack Project: openstack/tempest-lib  ca96243e586433a6b4aca769921462079ba67449
throw BadAltAuth exception when alt auth == auth

If we are expecting to replace a thing by switching to alternate auth,
but don't, this should be an exception. Otherwise it's easy to
obliviously not change to an alt auth element and be doing the wrong
thing.
Fixing and extending unit tests accordingly.

Change-Id: I738098c061c64d8bb68670ecd9ade616d3a9d5ab