Updated openstack/openstack Project: openstack-infra/project-config  e90c577968630526ff1ee6e90d7d59cd847ae388
Handle new release versions in doc publishing

OpenStack has stopped using date based versioning. Dates sort before
shorter semver'd versions which breaks the doc builds determination if
the docs should be writting to root of doc location or /$TAG. We can use
the old hack that we have in place for the clients on all the projects
now to fix this.

Change-Id: I71b0bb6604a9adf3cb898d0b3fdff77f900f83d4