Updated openstack/openstack Project: openstack-infra/puppet-drupal  9121247258bb79171eaa54ab6f8a8f225e48fc6e
Add reasonable default for $site_name parameter

If left undefined, the drupal class will create a vhost file named
50-.conf with an empty ServerName, and httpd will fail to start. Use
$::fqdn as a default site_name instead.

This will not affect Infra as Infra specifies this value explicitly.
Any downstream users must also be specifying this value explicitly for
the module to work, so this is a safe change.

Change-Id: Ia639f05a4ba7072812a1c4d6bb3a702fce493d02