Updated openstack/openstack Project: openstack-infra/puppet-drupal  c67a4231f8ae705c92b77ddc211e9ace8aa6379c
Fix ssl file logic

The default values for the $ssl_*_contents parameters is undef, so
checking whether they match empty string is not sensible. Fix the
conditionals to check for the case when the user has left the value
unset.

It is reasonable to conditionally not create these certificate files
even when $site_ssl_enabled is set because the path variables may be
set to a snakeoil cert.

Change-Id: I274dbe80a0547336601c12c4e876ba9591798996