Updated openstack/openstack Project: openstack-infra/project-config  905cf68fe3a49fcd73b55607f9ea7ec00e9798cf
puppet: retire beaker-upgrade jobs

beaker-upgrade jobs was really experimental and until now it never
really helped us in our CI & review process.

In the next cycle, we will push CI efforts to test OpenStack upgrades
with Puppet, but we don't know yet if it will be with Beaker or another
tool (our integration jobs are running in bash).

In the meantime, let's retire the jobs which are now useless.
They don't get attention and are quite unstable.

Change-Id: If2372fc3382d329ae7cff57b61f15f8d17083721