Updated openstack/openstack Project: openstack-infra/puppet-drupal  32eb6d3284b7fa403467d2a5dade7e37fe8f79d3
Fix cron setup logic

If $site_base_url and $conf are unset, catalog compilation fails
because the cron resource tries to access $conf['cron_key'] and $conf
is not a hash. If $conf is a hash but does not contain the 'cron_key'
key, a cron is created with the command

 wget -O /dev/null -q -t 1 false/cron.php?cron_key=

which is not sensible and will result in failed crons. It only makes
sense to create this cron if $site_base_url and $conf are set, so this
patch checks that they are set before creating the cron resource.

Change-Id: Ib40350e7b3c5c6261812d00a20decd8692acb4bb