Updated openstack/openstack Project: openstack-infra/project-config  08bfbe0ea5e62266aad252efb3a1f3212260ca87
Ignore non release versions when publishing latest docs

Ignore alpha, beta, rc versions when determining if we are publishing
the latest version of a projects documentation. These pre release
versions will still publish to /developer/project/$version but won't
update the root location of /developer/project.

This works around an issue with sort -V sorting the alphas, betas, and
rcs higher than the actual release. It also arguable better represents
what we want to do when publishing the latest of a release.

Change-Id: I2a4cd00414dee2e8307c681ed2bc6b61f7c670d6