Updated openstack/openstack Project: openstack/horizon  7143125e726c96d48d6409184e6bbb67d0c399d9
Merge tag '8.0.0'

This is a null-merge of the 8.0.0 release tag back into the master
branch so that the 8.0.0 tag will appear in the git commit history of
the master branch. It contains no actual changes to the master branch,
regardless of how our code review system's UI represents it. Please
ask in #openstack-infra if you have any questions, and otherwise try
to merge this as quickly as possible to avoid later conflicts on the
master branch.

Change-Id: I9e345994ed5c26792b86809000161a5384eaea28