Updated openstack/openstack Project: openstack-infra/puppet-kibana  304a9e6e81ec46be5b41a790552b2f567269094c
Remove trailing / from config.js url

We are sensitive to trailing slashes in the url for the elasticsearch
api due to the way kibana3 constructs its urls and how we proxy ES api.
Basically if ther config has a trailing / then we double up /'s and the
mod rewrite rules for rewriting the ES api breaks.

Simple fix is to have config.js remove and /s for us so that puppet
doesn't have to be too smart.

Fix here adapted from suggested at
https://stackoverflow.com/questions/6680825/return-string-without-trailing-slash

Change-Id: Ica46003f8a3706c88e9e2bf58e0d3c8ea7c4a50d