Updated openstack/openstack Project: openstack-infra/elastic-recheck  882a8c2afd3d607fe3a4a1c045901ae32d7bed99
Switch to kibana3 link generation

This makes use of the kibana3 logstash dashboard's query and from
parameters to link in queries similar to how we linked in kibana2
queries with base64 encoding. This however uses a saner urlencode
instead of base64 so you should be able to read them as a human.

Note we shouldn't land this until we have switched logstash.o.o over to
kibana3.

Change-Id: Idc1d637db7e1142a24adab2d21bc747d3002dd56
Depends-On: I176148c3bdaf603a1703d215ac2086101ecabe21