Updated openstack/openstack Project: openstack/ceilometer  5d40a1f9c88b9180c3d9686aa750c662a9e810f6
Merge tag '5.0.0'

This is a null-merge of the 5.0.0 release tag back into the master
branch so that the 5.0.0 tag will appear in the git commit history of
the master branch. It contains no actual changes to the master branch,
regardless of how our code review system's UI represents it. Please
ask in #openstack-infra if you have any questions, and otherwise try
to merge this as quickly as possible to avoid later conflicts on the
master branch.

Change-Id: Iac5daec1713c3251447d2353edbfe2a317fb19ef