Updated openstack/openstack Project: openstack-infra/devstack-gate  123eb1951b4c508dd57b2b0c58719e2750922bd1
Remove duplicates from the PROJECTS list

Duplicates in the PROJECTS list confuse workspace setup because the git
setup isn't idempotent. Easy fix for that is to remove duplicates and
only ever setup a repo once.

This will help us reduce the list of projects in the default PROJECTS
list because it will let us double account things during the transition.

Change-Id: I875641a5d00bc8afad97cbca6760ce66ac4a5c15