Updated openstack/openstack Project: openstack/heat  66c36f4ac8de36f1e1b59ac076e309e8958900cf
Merge tag '5.0.0'

This is a null-merge of the 5.0.0 release tag back into the master
branch so that the 5.0.0 tag will appear in the git commit history of
the master branch. It contains no actual changes to the master branch,
regardless of how our code review system's UI represents it. Please
ask in #openstack-infra if you have any questions, and otherwise try
to merge this as quickly as possible to avoid later conflicts on the
master branch.

Change-Id: Iccce01a30b9f1d1622c18834758e2470b936c7d9