Updated openstack/openstack Project: openstack/trove-integration  dfa3ae91ab64a487fb46e3c04755c0e811e804e7
Update redstack to Fedora 22

A recent change provided a refresh of the Fedora support
in redstack and its elements for F21. F22 is actually the
current release, the DIB default cloud image, and now has
packages for percona-xtrabackup. This change proposes to:

- make dnf (instead of yum) the new default Fedora package
manager for redstack and the Trove elements
- update mysql and mariadb elements to make use of the
Fedora distro version of percona-xtrabackup
- undo some unfortunate hacks needed in F21 relating to
pip and the python-six library
- miscellaneous SSH file permission fixes for Fedora

As before, this change in itself does not necessarily
provide complete and tested dev images for ALL the current
datastores, but instead focuses on mysql 5.6 and mariadb 5.5.

Now, the DIB_RELEASE will default to 22 and a kick-start can
be issued as:
MYSQL_HOST=127.0.0.1 DISTRO=fedora ./redstack kick-start mysql

Change-Id: I5f195ea7e8ab4d5b8e97801d82d0897469c811d2
Closes-Bug: 1501485