Updated openstack/openstack Project: openstack/trove  7d10b172f5de6227e096228aa3121f638d8f346f
Merge tag '4.0.0'

This is a null-merge of the 4.0.0 release tag back into the master
branch so that the 4.0.0 tag will appear in the git commit history of
the master branch. It contains no actual changes to the master branch,
regardless of how our code review system's UI represents it. Please
ask in #openstack-infra if you have any questions, and otherwise try
to merge this as quickly as possible to avoid later conflicts on the
master branch.

Change-Id: Ida15251e735add79f9b495ab8da14bd952e053b9