Updated openstack/openstack Project: openstack/swift  b59868f3533533897095779f224698ae4068c622
Interpolate the explanation string not whole HTML body

The only reason this exists is that I promised to do it.
But in our case, there's no big advantage, and here's why.

The general thinking goes that strings must be interpolated
first because the body may contain a syntax that confuses the
interpolation. So this patch makes the code "more correct".
However, our HTML template is tightly controlled. It's not
like it contains additional percents.

So I'll just leave this here for now while I'm asking if
the content type is set correctly.

Change-Id: Ia18aeb0f94ef389f8b95450986a566e5fa06aa10