Updated openstack/openstack Project: openstack/keystone  8a6c2846d0bf1239a2ebd37fd4905e862f3b60ee
switch to oslo.cache

Use oslo.cache now that it's been made available.

Note that we only deprecate the cache backends, but not remove
them, they are public APIs and subject to deprecation.

Also remove most of the caching code, but keep the portions
that register backends, since they are still available.

Note that the tests that were deleted were testing cache
functionality, they now exist in oslo.cache.

Co-Authored-By: Brant Knudson <bknudson@us.ibm.com>

Change-Id: I39d3ff1dd2edad79effa56450b444174b09f971e