Updated openstack/openstack Project: openstack/os-net-config  f885df39460d7d8d69e6a3375da69ceead42e6a7
Add Linux Bridge capability to os-net-config ifcfg

This patch adds support for Linux Bridges to os-net-config. This is
done completely with ifcfg files, brctl is not used directly.
Hierarchy is preserved, so a Linux Bridge may have a Linux Bond
as a member, which in turn may have multiple interfaces as members.
This changeset has been updated to include a more specific example
for Linux bridge configuration (that doesn't combine bridging and
bonding).

This change depends on the change to add support for Linux Bonds.

Change-Id: I1ddacd514b02af30139a868071d82cde19b1f946