Updated openstack/openstack Project: openstack/os-net-config  a7d9cc1129370f543ab05935bcff17896d09be70
Fix kwarg assignment in objects.py

This change modifies the kwargs in the object instantiation to
use immutable default assignments. This is for object safety,
since a mutable object will not point to a new memory address
on modification. In Python, lists and dicts should not be used
as default assignments for kwargs.

Change-Id: I73993df12ca317161ba32e02a98a5f40ef8904f3