Updated openstack/openstack Project: openstack-dev/devstack  0ba18481672964808bbbc4160643387dc931c654
Pin bashate and allow for substitution

I want to release a new bashate, but I also don't want to risk
consuming it before fully testing it.  By pinning here, we can run all
our usual CI on new versions before accepting them.

Additionally, allow substitution of the bashate dependency via an
environment variable.  We can use this in a bashate test to substitute
a path to a checkout with any changes applied.

Change-Id: I165c4d66db8b7bdcff235ef7d8c99029637bb76a