Updated openstack/openstack Project: openstack/neutron-vpnaas  4c6d44eb28508d4621c87198ba9e0d7d48f2c2c5
Merge tag '7.0.0'

This is a null-merge of the 7.0.0 release tag back into the master
branch so that the 7.0.0 tag will appear in the git commit history of
the master branch. It contains no actual changes to the master branch,
regardless of how our code review system's UI represents it. Please
ask in #openstack-infra if you have any questions, and otherwise try
to merge this as quickly as possible to avoid later conflicts on the
master branch.

Change-Id: I8c6c59fec5881d542643095dcfda3171875e42ba