Updated openstack/openstack Project: openstack-infra/shade  d60e11e51bcad850cfb29b242a727df547692f0c
Handle list_servers caching more directly

Some of the operations in shade are more expensive than others and need
a different approach to caching. For instance, nodepool is ALWAYS
building and deleting nodes, which means that any sort of
active-invalidation scheme renders the cache unusable, while a
time-based scheme that would make sense for flavors would be massively
too long.

This is the first stab but is really only going to be step one - and is
essentially the nodepool list_servers caching approach. list_servers is
cached by hand into a local mutex-protected list and is only updated
every five seconds.

Moving forward, the cache length needs to be configurable, but that
wants to be systematic rather than specific. Also, a local thread safe
dictionary dogpile cache should be written and we should use it by
default so that shared caches like redis or consul could be used in a
distributed system.

Change-Id: I430b4beaf3b05cb255c1cf5d180af450fd6429b2