Updated openstack/openstack Project: openstack/tripleo-heat-templates  0f2bf59910639c62442aa3a1266ea4e67d76d25a
Force stop a single node pacemaker on yum update

Currently package updates won't occur on a single node
non-HA pacemaker managed Controller because stopping
the node loses the quorum of 1.

This change gets the count of current nodes in the cluster and
if the count is 1 then specify --force when doing a pcs cluster stop.

Change-Id: I0de2488e24f1ef53a935dbc90ec6de6142bb4264