Updated openstack/openstack Project: openstack/barbican  6bef9eb964b6f8ae9c4a42c057e313c36d9f41d6
Merge tag '1.0.0'

This is a null-merge of the 1.0.0 release tag back into the master
branch so that the 1.0.0 tag will appear in the git commit history of
the master branch. It contains no actual changes to the master branch,
regardless of how our code review system's UI represents it. Please
ask in #openstack-infra if you have any questions, and otherwise try
to merge this as quickly as possible to avoid later conflicts on the
master branch.

Change-Id: I42217149f3be2db4a0278ef485c42d6555397f15