Updated openstack/openstack Project: openstack/glance  809d88c75c13c455747245a1f8d62a7939bf8c2f
Show the file name when there is an error loading an image metadef file

New image metadata can be added by creating a new json file in the directory
glance/etc/metadefs (this metadata shows up in the image metadata UI on
Horizon dashboard). If the file is loaded successfully, a log message like
the following is printed:

2015-08-09 11:48:34.713 INFO glance.db.sqlalchemy.metadata [-]
File /etc/glance/metadefs/operating-system.json loaded to database.

However, if there is an error loading the file, we get a log message like:

2015-08-09 11:48:34.875 ERROR glance.db.sqlalchemy.metadata [-]
Expecting , delimiter: line 15 column 9 (char 579)

In other words, the file name is not printed. Printing the file name in the
error message will make it easier to debug as we can search the logs using
the file name.

Change-Id: If72fdbb00c58e42aa128ab2b4461a2ae8e55c5aa
Closes-Bug: #1483049